Passive match notification & Deduplication fix
Added a label per trade which shows that the trade is either "matched"
or "waiting to be matched".
Matched trades are not deletable to serve as a note. Matched trade are
not considered in the trade integrity contraints set or for the
packages generation.

Fixed: deleting a matched trade gives you back the cards.